+++ src/vscode-bicep/src/tree/AzLocationTreeItem.ts
