+++ src/vscode-bicep/src/tree/GenericAzExtTreeItem.ts
