+++ src/vscode-bicep/src/tree/ResourceGroupTreeItem.ts
@@ -21,9 +21,9 @@ import {
   nonNullProp,
 } from "@microsoft/vscode-azext-utils";
 
-import { createResourceManagementClient } from "../../azure/azureClients";
-import { localize } from "../../utils/localize";
-import { OutputChannelManager } from "../../utils/OutputChannelManager";
+import { createResourceManagementClient } from "../azure/azureClients";
+import { localize } from "../utils/localize";
+import { OutputChannelManager } from "../utils/OutputChannelManager";
 import { GenericAzExtTreeItem } from "./GenericAzExtTreeItem";
 
 // Represents an Azure subscription. Used to display resource groups related to the subscription
